move loggables to domain
