+++ apps/server/src/modules/user-login-migration/api/user-login-migration.controller.ts
@@ -19,7 +19,7 @@ import {
 	UserLoginMigrationAlreadyClosedLoggableException,
 	UserLoginMigrationGracePeriodExpiredLoggableException,
 	UserLoginMigrationNotFoundLoggableException,
-} from '../loggable';
+} from '../domain/loggable';
 import {
 	CloseMigrationWizardUc,
 	CloseUserLoginMigrationUc,
