+++ apps/server/src/modules/user-login-migration/domain/loggable/debug/school-migration-successful.loggable.spec.ts
