+++ apps/server/src/modules/user-login-migration/domain/loggable/external-school-number-missing.loggable-exception.ts
