+++ apps/server/src/modules/user-login-migration/domain/loggable/invalid-user-login-migration.loggable-exception.ts
