+++ apps/server/src/modules/user-login-migration/domain/loggable/moin-schule-system-not-found.loggable-exception.ts
