+++ apps/server/src/modules/user-login-migration/domain/loggable/school-migration-database-operation-failed.loggable-exception.spec.ts
