+++ apps/server/src/modules/user-login-migration/domain/loggable/school-number-mismatch.loggable-exception.ts
