+++ apps/server/src/modules/user-login-migration/domain/loggable/school-number-missing.loggable-exception.spec.ts
