+++ apps/server/src/modules/user-login-migration/domain/loggable/user-login-migration-already-closed.loggable-exception.ts
