+++ apps/server/src/modules/user-login-migration/domain/loggable/user-login-migration-grace-period-expired-loggable.exception.ts
