+++ apps/server/src/modules/user-login-migration/domain/loggable/user-login-migration-invalid-admin.loggable-exception.ts
