+++ apps/server/src/modules/user-login-migration/domain/loggable/user-login-migration-invalid-external-school-id.loggable-exception.spec.ts
