+++ apps/server/src/modules/user-login-migration/domain/loggable/user-login-migration-mandatory.loggable.spec.ts
