+++ apps/server/src/modules/user-login-migration/domain/loggable/user-login-migration-not-found.loggable-exception.ts
