+++ apps/server/src/modules/user-login-migration/domain/loggable/user-login-migration-school-already-migrated.loggable-exception.spec.ts
