+++ apps/server/src/modules/user-login-migration/domain/loggable/user-login-migration-user-already-migrated.loggable-exception.ts
