+++ apps/server/src/modules/user-login-migration/domain/loggable/user-migration-database-operation-failed.loggable-exception.ts
