+++ apps/server/src/modules/user-login-migration/domain/loggable/user-not-migrated.loggable-exception.spec.ts
